Break up _cookbook.py
